chore: refactor patterns
